Abstract waitUntil
