+++ lib/index.js
@@ -4,23 +4,7 @@ import { fileURLToPath } from 'url';
 import net from 'net';
 import genericPool from 'generic-pool';
 import finalStream from 'final-stream';
-
-const waitUntil = (condition) => {
-  if (condition()) {
-    return Promise.resolve();
-  }
-
-  return new Promise((resolve) => {
-    const interval = setInterval(() => {
-      if (!condition()) {
-        return;
-      }
-
-      clearInterval(interval);
-      resolve();
-    }, 0);
-  });
-};
+import waitUntil from './waitUntil.js';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
