+++ lib/waitUntil.js
@@ -1,0 +1,18 @@
+const waitUntil = (condition) => {
+  if (condition()) {
+    return Promise.resolve();
+  }
+
+  return new Promise((resolve) => {
+    const interval = setInterval(() => {
+      if (!condition()) {
+        return;
+      }
+
+      clearInterval(interval);
+      resolve();
+    }, 0);
+  });
+};
+
+export default waitUntil;
